feat(blackjack): remove old logic and replace with new logic
