+++ src/main/java/blackjack/Application.java
@@ -1,15 +1,15 @@
 package blackjack;
 
-import blackjack.controller.BlackjackController2;
-import blackjack.domain.betting2.PlayerBettings;
-import blackjack.domain.BlackjackGame2;
+import blackjack.controller.BlackjackController;
+import blackjack.domain.betting.PlayerBettings;
+import blackjack.domain.BlackjackGame;
 
 public class Application {
 
-    private static final BlackjackController2 controller = new BlackjackController2();
+    private static final BlackjackController controller = new BlackjackController();
 
     public static void main(String[] args) {
-        final BlackjackGame2 game = controller.initializeGame();
+        final BlackjackGame game = controller.initializeGame();
         final PlayerBettings bettings = controller.initializeBettings(game);
 
         controller.playGame(game);
