+++ src/main/java/blackjack/domain/betting/BettingResults.java
@@ -11,14 +11,14 @@ public class BettingResults {
     private final List<BettingResult> value;
 
     public BettingResults(final Participant dealer, final PlayerBettings playerBettings) {
-        List<BettingResult> playerBettingResults = initAllPlayerBettingResults(playerBettings, dealer);
-        BettingResult dealerBettingResult = initDealerResultFrom(dealer, playerBettingResults);
+        List<BettingResult> playerBettingResults = initAllPlayerResults(playerBettings, dealer);
+        BettingResult dealerBettingResult = initDealerResult(dealer, playerBettingResults);
 
-        this.value = getUnmodifiableOrderedListOf(dealerBettingResult, playerBettingResults);
+        this.value = toList(dealerBettingResult, playerBettingResults);
     }
 
-    private List<BettingResult> getUnmodifiableOrderedListOf(final BettingResult dealerBettingResult,
-                                                             final List<BettingResult> playerBettingResults) {
+    private List<BettingResult> toList(final BettingResult dealerBettingResult,
+                                       final List<BettingResult> playerBettingResults) {
         List<BettingResult> bettingResults = new ArrayList<>();
 
         bettingResults.add(dealerBettingResult);
@@ -27,32 +27,33 @@ public class BettingResults {
         return Collections.unmodifiableList(bettingResults);
     }
 
-    private List<BettingResult> initAllPlayerBettingResults(final PlayerBettings playerBettings,
-                                                            final Participant dealer) {
+    private List<BettingResult> initAllPlayerResults(final PlayerBettings playerBettings,
+                                                     final Participant dealer) {
         return playerBettings.getValue()
                 .stream()
-                .map(playerBetting -> initPlayerBettingResult(playerBetting, dealer))
+                .map(playerBetting -> initPlayerResult(playerBetting, dealer))
                 .collect(Collectors.toUnmodifiableList());
     }
 
-    private BettingResult initPlayerBettingResult(final PlayerBetting playerBetting, final Participant dealer) {
+    private BettingResult initPlayerResult(final PlayerBetting playerBetting,
+                                           final Participant dealer) {
         final Participant player = playerBetting.getPlayer();
-        final int profit = playerBetting.getProfit(dealer);
+        final int profit = playerBetting.profit(dealer);
 
         return new BettingResult(player, profit);
     }
 
-    private BettingResult initDealerResultFrom(final Participant dealer,
-                                               final List<BettingResult> playerBettingResults) {
-        final int totalPlayerProfit = getProfitSumOf(playerBettingResults);
+    private BettingResult initDealerResult(final Participant dealer,
+                                           final List<BettingResult> playerBettingResults) {
+        final int totalPlayerProfit = profitSumOf(playerBettingResults);
         final int dealerProfit = totalPlayerProfit * -1;
 
         return new BettingResult(dealer, dealerProfit);
     }
 
-    private int getProfitSumOf(List<BettingResult> playerBettingResult) {
+    private int profitSumOf(List<BettingResult> playerBettingResult) {
         return playerBettingResult.stream()
-                .mapToInt(BettingResult::getMoneyOutcome)
+                .mapToInt(BettingResult::getProfit)
                 .sum();
     }
