+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,61 +1,45 @@
 package blackjack.domain.participant;
 
+import static blackjack.validator.PlayerNameValidator.validateNameNotBlank;
+import static blackjack.validator.PlayerNameValidator.validateNameNotDealer;
+
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardBundle;
+import blackjack.domain.hand.CardHand;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class Player extends Participant {
+public final class Player extends BlackjackParticipant {
 
-    private static final int INITIAL_PLAYER_OPEN_CARDS_COUNT = 2;
-    private static final String BLACK_NAME_INPUT_EXCEPTION_MESSAGE = "플레이어는 이름을 지녀야 합니다.";
-    private static final String INVALID_PLAYER_NAME_EXCEPTION_MESSAGE = "플레이어의 이름은 딜러가 될 수 없습니다.";
+    private static final int OPEN_CARDS_COUNT = 2;
 
     private final String name;
 
-    private Player(final String name, final CardBundle cardBundle) {
-        super(cardBundle, Player::stayStrategy);
+    public Player(final String name, final CardHand cardHand) {
+        super(cardHand);
+        validateNameNotBlank(name);
+        validateNameNotDealer(name);
         this.name = name;
     }
 
-    private static boolean stayStrategy(final CardBundle cardBundle) {
+    @Override
+    public List<Card> openInitialCards() {
+        return cardHand.getCardBundle()
+                .getCards()
+                .stream()
+                .limit(OPEN_CARDS_COUNT)
+                .collect(Collectors.toUnmodifiableList());
+    }
+
+    @Override
+    protected boolean shouldStay() {
+        CardBundle cardBundle = cardHand.getCardBundle();
         return cardBundle.isBlackJackScore();
-    }
-
-    public static Player of(final String name, final CardBundle cardBundle) {
-        validateName(name);
-        return new Player(name, cardBundle);
-    }
-
-    private static void validateName(final String name) {
-        validateNameExistence(name);
-        validateNameNotDealer(name);
-    }
-
-    private static void validateNameExistence(final String name) {
-        if (name.isBlank()) {
-            throw new IllegalArgumentException(BLACK_NAME_INPUT_EXCEPTION_MESSAGE);
-        }
-    }
-
-    private static void validateNameNotDealer(final String name) {
-        if (name.equals(Dealer.UNIQUE_NAME)) {
-            throw new IllegalArgumentException(INVALID_PLAYER_NAME_EXCEPTION_MESSAGE);
-        }
     }
 
     @Override
     public String getName() {
         return name;
-    }
-
-    @Override
-    public List<Card> getInitialOpenCards() {
-        return cardHand.getCardBundle()
-                .getCards()
-                .stream()
-                .limit(INITIAL_PLAYER_OPEN_CARDS_COUNT)
-                .collect(Collectors.toUnmodifiableList());
     }
 
     @Override
