--- src/main/java/blackjack/domain/state/CardHand.java
@@ -1,21 +1,0 @@
-package blackjack.domain.state;
-
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardBundle;
-
-public interface CardHand {
-
-    CardHand hit(Card card);
-
-    CardHand stay();
-
-    boolean isFinished();
-
-    boolean isBlackjack();
-
-    boolean isBust();
-
-    double getBettingYieldVersus(CardHand targetState);
-
-    CardBundle getCardBundle();
-}
