--- src/main/java/blackjack/strategy/CardsViewStrategy2.java
@@ -1,9 +1,0 @@
-package blackjack.strategy;
-
-import blackjack.domain.participant2.Participant;
-
-@FunctionalInterface
-public interface CardsViewStrategy2 {
-
-    void print(Participant participant);
-}
