+++ src/test/java/blackjack/domain/BlackjackGameTest.java
@@ -1,47 +1,35 @@
 package blackjack.domain;
 
-import static blackjack.fixture.CardBundleFixture.BLACKJACK_CARD_BUNDLE;
-import static blackjack.fixture.CardBundleFixture.CARD_BUNDLE_16;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.card.CardBundle;
 import blackjack.domain.card.CardDeck;
-import blackjack.domain.participant.Dealer;
+import blackjack.domain.hand.CardHand;
 import blackjack.domain.participant.Participant;
-import blackjack.domain.participant.Player;
-import blackjack.strategy.CardBundleStrategy;
 import blackjack.strategy.CardsViewStrategy;
 import blackjack.strategy.HitOrStayChoiceStrategy;
 import java.util.List;
+import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 public class BlackjackGameTest {
 
-    private static final List<String> PLAYER_NAMES_LIST = List.of("player1", "player2");
-
-    private static final CardBundleStrategy prodStrategy =
-            (cardStack) -> CardBundle.of(cardStack.pop(), cardStack.pop());
-    private static final CardBundleStrategy cardBundleOfSixteenStrategy = (cardStack) -> CARD_BUNDLE_16();
-    private static final CardBundleStrategy cardBundleOfBlackjackStrategy = (cardStack) -> BLACKJACK_CARD_BUNDLE();
-
-    private static final HitOrStayChoiceStrategy DRAW_CHOICE = () -> true;
-    private static final CardsViewStrategy CARDS_VIEW_STRATEGY = (dealer) -> {
+    private static final HitOrStayChoiceStrategy HIT_CHOICE = () -> true;
+    private static final CardsViewStrategy VIEW_STRATEGY = (dealer) -> {
     };
 
     @DisplayName("생성자 테스트")
     @Nested
     class ConstructorTest {
 
-        @DisplayName("생성자는 1명 이상의 플레이어명을 리스트로 받아 게임을 생성한다.")
+        @DisplayName("생성자는 1명 이상의 플레이어명을 순서대로 받아 게임을 생성한다.")
         @Test
-        void constructor_initsGameWithPlayerNames() {
-            BlackjackGame blackjackGame = new BlackjackGame(
-                    new CardDeck(), List.of("hudi", "jeong"), prodStrategy);
+        void constructor() {
+            BlackjackGame blackjackGame = new BlackjackGame(new CardDeck(), List.of("hudi", "jeong"));
 
-            List<Participant> participants = blackjackGame.getParticipants().getPlayers();
+            List<Participant> participants = extractPlayers(blackjackGame);
 
             assertThat(participants.size()).isEqualTo(2);
             assertThat(participants.get(0).getName()).isEqualTo("hudi");
@@ -51,7 +39,7 @@ public class BlackjackGameTest {
         @DisplayName("생성자 파라미터에 들어오는 플레이어명 리스트가 비어있으면 예외가 발생한다.")
         @Test
         void constructor_throwsExceptionOnNoPlayerNameInput() {
-            assertThatThrownBy(() -> new BlackjackGame(new CardDeck(), List.of(), prodStrategy))
+            assertThatThrownBy(() -> new BlackjackGame(new CardDeck(), List.of()))
                     .isInstanceOf(IllegalArgumentException.class)
                     .hasMessage("플레이어가 없는 게임은 존재할 수 없습니다.");
         }
@@ -59,66 +47,48 @@ public class BlackjackGameTest {
         @DisplayName("생성자 파라미터에 들어오는 플레이어명들 중 중복이 존재하면 예외가 발생한다.")
         @Test
         void constructor_throwsExceptionOnDuplicatePlayerNameInput() {
-            assertThatThrownBy(() -> new BlackjackGame(
-                    new CardDeck(), List.of("중복", "중복"), prodStrategy))
+            assertThatThrownBy(() -> new BlackjackGame(new CardDeck(), List.of("중복", "중복")))
                     .isInstanceOf(IllegalArgumentException.class)
                     .hasMessage("플레이어명은 중복될 수 없습니다.");
         }
     }
 
-    @DisplayName("isBlackjackDealer 메서드는 Dealer 인스턴스의 isBlackjack 메서드를 호출한다.")
+    @DisplayName("drawAllPlayerCards 메서드를 호출하면 모든 플레이어는 더 이상 hit이 불가능한 상태가 된다.")
     @Test
-    void isBlackjackDealer_trueOnDealerBlackjack() {
-        BlackjackGame blackjackGame = new BlackjackGame(
-                new CardDeck(), PLAYER_NAMES_LIST, cardBundleOfBlackjackStrategy);
-
-        boolean actual = blackjackGame.isBlackjackDealer();
-
-        assertThat(actual).isTrue();
-    }
-
-    @DisplayName("getParticipants 메서드는 Dealer 인스턴스와 1명 이상의 Player 인스턴스로 구성된 컬렉션을 반환한다.")
-    @Test
-    void getParticipants_returnsListOfParticipantsWithDealerAtIndexZero() {
-        BlackjackGame blackjackGame = new BlackjackGame(
-                new CardDeck(), List.of("p1", "p2", "p3"), cardBundleOfSixteenStrategy);
-
-        List<Participant> actual = blackjackGame.getParticipants().getValue();
-
-        assertThat(actual.get(0)).isInstanceOf(Dealer.class);
-        for (int i = 1; i < actual.size(); i++) {
-            assertThat(actual.get(i)).isInstanceOf(Player.class);
-        }
-    }
-
-    @DisplayName("drawAllPlayerCards 메서드 호출 후 모든 플레이어는 더 이상 드로우를 할 수 없게 된다.")
-    @Test
-    void drawAllPlayerCards() {
+    void drawAllPlayerCards_isFinishedAfterKeepHitting() {
         for (int i = 0; i < 1000; i++) {
             BlackjackGame blackjackGame = new BlackjackGame(
-                    new CardDeck(), List.of("p1", "p2", "p3"), prodStrategy);
+                    new CardDeck(), List.of("p1", "p2", "p3"));
 
-            blackjackGame.drawAllPlayerCards(DRAW_CHOICE, CARDS_VIEW_STRATEGY);
-            blackjackGame.getParticipants()
-                    .getPlayers()
-                    .stream()
-                    .map(Participant::canDraw)
-                    .forEach(playerCanDraw -> assertThat(playerCanDraw).isFalse());
+            blackjackGame.drawAllPlayerCards(HIT_CHOICE, VIEW_STRATEGY);
+            List<CardHand> playerHands = extractPlayers(blackjackGame).stream()
+                    .map(Participant::getHand)
+                    .collect(Collectors.toUnmodifiableList());
+
+            for (CardHand playerHand : playerHands) {
+                assertThat(playerHand.isFinished()).isTrue();
+            }
         }
     }
 
-    @DisplayName("drawDealerCards 메서드 호출 후 딜러는 더 이상 드로우를 할 수 없게 된다.")
+    @DisplayName("drawDealerCards 메서드 호출 후 딜러는 더 이상 hit이 불가능한 상태가 된다.")
     @Test
-    void drawDealerCards() {
+    void drawDealerCards_isFinishedAfterKeepHitting() {
         for (int i = 0; i < 1000; i++) {
-            BlackjackGame blackjackGame = new BlackjackGame(
-                    new CardDeck(), List.of("p1", "p2", "p3"), prodStrategy);
+            BlackjackGame blackjackGame = new BlackjackGame(new CardDeck(), List.of("p1"));
+            blackjackGame.drawDealerCards(VIEW_STRATEGY);
 
-            blackjackGame.drawDealerCards(CARDS_VIEW_STRATEGY);
-            Participant dealer = blackjackGame.getParticipants().getDealer();
+            CardHand dealerHand = extractDealer(blackjackGame).getHand();
 
-            boolean actual = dealer.canDraw();
-            assertThat(actual).isFalse();
+            assertThat(dealerHand.isFinished()).isTrue();
         }
+    }
+
+    private Participant extractDealer(BlackjackGame blackjackGame) {
+        return blackjackGame.getParticipants().getDealer();
+    }
+
+    private List<Participant> extractPlayers(BlackjackGame blackjackGame) {
+        return blackjackGame.getParticipants().getPlayers();
     }
 }
